+++ sickbeard/providers/grabtheinfo.py
@@ -34,8 +34,7 @@ class GrabTheInfoProvider(generic.TorrentProvider):
         self.urls = {'config_provider_home_uri': self.url_base,
                      'login': self.url_base + 'rules.php',
                      'browse': self.url_base + 'browse.php?%s&incldead=%s&blah=0%s',
-                     'search': '&search=%s',
-                     'get': self.url_base + '%s'}
+                     'search': '&search=%s'}
 
         self.categories = {'shows': [36, 32, 43, 56, 8, 65, 61, 10]}